@@ -1,4 +1,3 @@
-using System;
 using System.Globalization;
 
 namespace F0.Minesweeper.Components.Logic.Cell
@@ -23,7 +22,7 @@ namespace F0.Minesweeper.Components.Logic.Cell
 		{
 			if (activeTranslation is null && adjacentMineCount is null)
 			{
-				throw new ArgumentNullException(nameof(adjacentMineCount));
+				ArgumentNullException.ThrowIfNull(adjacentMineCount);
 			}
 
 			if (activeTranslation.HasValue)
