@@ -1,6 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 using F0.Minesweeper.Components.Abstractions;
 using F0.Minesweeper.Components.Abstractions.Enums;
 using F0.Minesweeper.Components.Logic.Cell;
@@ -12,6 +9,7 @@ namespace F0.Minesweeper.Components.Pages.Game.Modules
 	public partial class Cell : ICell
 	{
 		[Parameter]
+		[EditorRequired]
 		public Location? Location { get; set; }
 
 		public Func<Location, Task>? UncoveredAsync { get; set; }
@@ -59,10 +57,7 @@ namespace F0.Minesweeper.Components.Pages.Game.Modules
 
 		protected override void OnParametersSet()
 		{
-			if (Location is null)
-			{
-				throw new ArgumentNullException(nameof(Location));
-			}
+			ArgumentNullException.ThrowIfNull(Location);
 		}
 
 		public void SetUncoveredStatus(CellInteractionType cellInteraction, bool isMine, byte adjacentMineCount)
@@ -108,7 +103,7 @@ namespace F0.Minesweeper.Components.Pages.Game.Modules
 
 		private void OnRightClick()
 		{
-			TryUpdateStatus(CellInteractionType.RightClick);
+			_ = TryUpdateStatus(CellInteractionType.RightClick);
 		}
 
 		private bool TryUpdateStatus(CellInteractionType inputCommand, bool? isMine = null, byte? adjacentMineCount = null)
