@@ -1,5 +1,3 @@
-using System;
-using System.Collections.Generic;
 using System.ComponentModel;
 using F0.Minesweeper.Logic.Abstractions;
 using F0.Minesweeper.Logic.LocationShuffler;
@@ -14,7 +12,7 @@ namespace F0.Minesweeper.Logic
 
 		public IMinefield Create(MinefieldOptions options)
 		{
-			_ = options ?? throw new ArgumentNullException(nameof(options));
+			ArgumentNullException.ThrowIfNull(options);
 
 			ILocationShuffler locationShuffler;
 
