@@ -2,7 +2,7 @@ using F0.Minesweeper.Components.Abstractions.Enums;
 
 namespace F0.Minesweeper.Components.Logic.Cell
 {
-	internal record CellStatusTransitionCommand
+	internal record class CellStatusTransitionCommand
 	{
 		internal CellInteractionType MouseButtonType { get; }
 		internal bool? IsMine { get; }
