@@ -1,13 +1,9 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
 using F0.Minesweeper.Components.Abstractions;
 using F0.Minesweeper.Components.Abstractions.Enums;
 using F0.Minesweeper.Components.Events;
 using F0.Minesweeper.Components.Logic.Game;
 using F0.Minesweeper.Logic.Abstractions;
-using Moq;
 using Prism.Events;
-using Xunit;
 
 namespace F0.Minesweeper.Components.Tests.Logic.Game
 {
@@ -61,7 +57,7 @@ namespace F0.Minesweeper.Components.Tests.Logic.Game
 			}
 
 			internal Task OnUpdateAsyncExposed(IEnumerable<UncoverableCell> uncoverableCells, Minesweeper.Logic.Abstractions.Location clickedLocation)
-				=> this.OnUpdateAsync(uncoverableCells, clickedLocation);
+				=> OnUpdateAsync(uncoverableCells, clickedLocation);
 		}
 	}
 }
