@@ -1,6 +1,4 @@
 using F0.Minesweeper.Logic.Abstractions;
-using FluentAssertions;
-using Xunit;
 
 namespace F0.Minesweeper.Logic.Tests
 {
@@ -27,7 +25,7 @@ namespace F0.Minesweeper.Logic.Tests
 		[Fact]
 		public void NewLocation_MaxIntParameterOnX()
 		{
-			uint testValue = int.MaxValue;
+			uint testValue = Int32.MaxValue;
 			Location locationUnderTest = new(testValue, 7);
 
 			locationUnderTest.X.Should().Be(testValue);
@@ -36,7 +34,7 @@ namespace F0.Minesweeper.Logic.Tests
 		[Fact]
 		public void NewLocation_MaxIntParameterOnY()
 		{
-			uint testValue = int.MaxValue;
+			uint testValue = Int32.MaxValue;
 			Location locationUnderTest = new(5, testValue);
 
 			locationUnderTest.Y.Should().Be(testValue);
